Adds TURN server, restructures the backend and adds CI for backend and frontend
