+++ backend/alembic/env.py
@@ -8,9 +8,9 @@ from sqlalchemy.ext.asyncio import async_engine_from_config
 from alembic import context
 
 # Import models and config
-from config import settings
-from database import Base
-from models import Meeting, Participant, MeetingLog  # Import all models
+from src.core.config import settings
+from src.db.database import Base
+from src.db.models import Meeting, Participant, MeetingLog  # Import all models
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
