+++ backend/pytest.ini
@@ -1,0 +1,15 @@
+[pytest]
+# Pytest configuration for async tests
+asyncio_mode = auto
+testpaths = tests
+python_files = test_*.py
+python_classes = Test*
+python_functions = test_*
+
+# Disable warnings
+filterwarnings =
+    ignore::DeprecationWarning
+    ignore::pytest.PytestUnraisableExceptionWarning
+
+# Show test output
+addopts = -v --tb=short
