+++ backend/src/__init__.py
@@ -1,0 +1,3 @@
+"""LinkUp Video Call Backend Application"""
+
+__version__ = "2.0.0"
