+++ backend/src/core/config.py
@@ -44,4 +44,3 @@ class Settings(BaseSettings):
 
 
 settings = Settings()
-
