+++ backend/tests/conftest.py
@@ -1,0 +1,62 @@
+import pytest
+import pytest_asyncio
+from httpx import AsyncClient
+from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
+from sqlalchemy.orm import sessionmaker
+
+from src.main import app
+from src.db.database import Base, get_db
+from src.core.config import settings
+
+# Test database URL (use in-memory SQLite for testing)
+TEST_DATABASE_URL = "sqlite+aiosqlite:///:memory:"
+
+# Create test engine
+test_engine = create_async_engine(TEST_DATABASE_URL, echo=False)
+
+# Create test session factory
+TestSessionLocal = sessionmaker(
+    test_engine,
+    class_=AsyncSession,
+    expire_on_commit=False,
+    autocommit=False,
+    autoflush=False,
+)
+
+
+@pytest_asyncio.fixture
+async def db_session():
+    """Create a fresh database session for each test"""
+    async with test_engine.begin() as conn:
+        await conn.run_sync(Base.metadata.create_all)
+    
+    async with TestSessionLocal() as session:
+        yield session
+    
+    async with test_engine.begin() as conn:
+        await conn.run_sync(Base.metadata.drop_all)
+
+
+@pytest_asyncio.fixture
+async def client(db_session):
+    """Create a test client with database override"""
+    from httpx import ASGITransport
+    
+    async def override_get_db():
+        yield db_session
+    
+    app.dependency_overrides[get_db] = override_get_db
+    
+    async with AsyncClient(transport=ASGITransport(app=app), base_url="http://test") as ac:
+        yield ac
+    
+    app.dependency_overrides.clear()
+
+
+@pytest.fixture
+def sample_meeting_data():
+    """Sample meeting data for tests"""
+    return {
+        "title": "Test Meeting",
+        "max_participants": 10
+    }
