+++ frontend/src/config.ts
@@ -2,17 +2,31 @@
 export const API_BASE_URL = import.meta.env.VITE_API_URL || 'http://localhost:8000';
 export const WS_BASE_URL = import.meta.env.VITE_WS_URL || 'ws://localhost:8000';
 
-// WebRTC Configuration with Google STUN servers
+// WebRTC Configuration with STUN and TURN servers
 export const RTC_CONFIG: RTCConfiguration = {
   iceServers: [
     {
-      urls: [
-        'stun:stun.l.google.com:19302',
-        'stun:stun1.l.google.com:19302',
-        'stun:stun2.l.google.com:19302',
-        'stun:stun3.l.google.com:19302',
-        'stun:stun4.l.google.com:19302',
-      ]
+      urls: 'stun:stun.relay.metered.ca:80'
+    },
+    {
+      urls: 'turn:global.relay.metered.ca:80',
+      username: '81df41330ffff0c3d75e77ed',
+      credential: 'yn1jtZhMFeVgGmYo'
+    },
+    {
+      urls: 'turn:global.relay.metered.ca:80?transport=tcp',
+      username: '81df41330ffff0c3d75e77ed',
+      credential: 'yn1jtZhMFeVgGmYo'
+    },
+    {
+      urls: 'turn:global.relay.metered.ca:443',
+      username: '81df41330ffff0c3d75e77ed',
+      credential: 'yn1jtZhMFeVgGmYo'
+    },
+    {
+      urls: 'turns:global.relay.metered.ca:443?transport=tcp',
+      username: '81df41330ffff0c3d75e77ed',
+      credential: 'yn1jtZhMFeVgGmYo'
     }
   ]
 };
